# Server-Side/Static/Swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "DouneStore API",
    "description": "API documentation for DouneStore",
    "version": "1.0.0"
  },
  "host": "localhost:5000",
  "basePath": "/",
  "schemes": ["http"],
  "paths": {
    "/register": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Register a new user",
        "description": "Register a new user and send OTP verification email",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "User registration details",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["email", "password", "fullname"],
              "properties": {
                "email": {
                  "type": "string",
                  "example": "user@example.com"
                },
                "password": {
                  "type": "string",
                  "example": "password123"
                },
                "fullname": {
                  "type": "string",
                  "example": "John Doe"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful registration request",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": true
                },
                "message": {
                  "type": "string",
                  "example": "OTP đã được gửi đến email của bạn"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": false
                },
                "message": {
                  "type": "string",
                  "example": "Error message"
                }
              }
            }
          }
        }
      }
    },
    "/verify": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Verify OTP and complete registration",
        "description": "Verify OTP sent to email and create new user account",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "OTP verification details",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["email", "otp"],
              "properties": {
                "email": {
                  "type": "string",
                  "example": "user@example.com"
                },
                "otp": {
                  "type": "string",
                  "example": "123456"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful verification",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": true
                },
                "message": {
                  "type": "string",
                  "example": "Đăng ký thành công"
                }
              }
            }
          },
          "400": {
            "description": "Invalid OTP or expired",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": false
                },
                "message": {
                  "type": "string",
                  "example": "OTP không hợp lệ"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": false
                },
                "message": {
                  "type": "string",
                  "example": "Error message"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}